moved tensor into core
