+++ hub/__init__.py
@@ -11,13 +11,12 @@ __pdoc__ = {
 }
 
 from .api.dataset_handler import dataset
-from .api.tensor import Tensor
 from .api.read import read
 from .util.bugout_reporter import hub_reporter
 
 load = dataset.load
 empty = dataset.empty
-__all__ = ["dataset", "Tensor", "read", "__version__", "load", "empty"]
+__all__ = ["dataset", "read", "__version__", "load", "empty"]
 
 __version__ = "2.0.2"
 __encoded_version__ = np.array(__version__)
