+++ hub/api/tests/test_api_with_compression.py
@@ -4,7 +4,7 @@ from hub.util.exceptions import (
     UnsupportedCompressionError,
 )
 import pytest
-from hub.api.tensor import Tensor
+from hub.core.tensor import Tensor
 from hub.tests.common import TENSOR_KEY
 from hub.tests.dataset_fixtures import enabled_datasets
 import numpy as np
